Implement asynchronous document structure extraction and enhance retrieval endpoints
- Added a new endpoint to start asynchronous document structure extraction, incorporating rate limiting and user access validation.
- Enhanced the existing document structure retrieval endpoint to support both tree and flat formats, with appropriate error handling for missing structures.
- Introduced a new asynchronous method in DocumentStructureService for extracting and aligning document structures, updating job statuses throughout the process.
- Updated the RateLimitService to ensure proper handling of rate limit exceptions.
- Refactored related services and controllers for improved organization and maintainability, including the addition of logging for better traceability.
- Added unit tests to validate the new functionalities and ensure robust interactions within the document processing services.

+++ src/main/java/uk/gegc/quizmaker/shared/rate_limit/RateLimitService.java
@@ -32,7 +32,7 @@ public class RateLimitService {
             int count = requestCount.getOrDefault(rateLimitKey, 0);
             if (count >= limitPerMinute) {
                 long retryAfter = java.time.Duration.between(now, firstInWindow.plusMinutes(1)).getSeconds();
-                throw new RateLimitExceededException("Too many requests for " + operation, retryAfter);
+                throw new RateLimitExceededException("Too many requests for " + operation, (int) retryAfter);
             }
             requestCount.put(rateLimitKey, count + 1);
         } else {
